requirements.txt
python-dotenv==1.1.1
pytz==2025.2
pdfplumber==0.11.7
requests==2.32.5
jupyterlab==4.4.9
streamlit==1.50.0
openai
tiktoken==0.11.0
anthropic==0.69.0
google-genai==1.41.0
llamaapi==0.1.36
xai-sdk==1.2.0
notion==0.0.28
notion-client==2.5.0
line-bot-sdk==3.19.1
mecab-python3==1.0.10
scikit-learn==1.7.2
wordcloud==1.9.4
openpyxl==3.1.5
chromadb==1.1.1
fastapi==0.118.0
uvicorn==0.37.0
gunicorn==23.0.0
pydantic==2.11.10
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
bcrypt==5.0.0
matplotlib==3.10.6
numpy==2.2.6
pandas==2.3.3
PyYAML==6.0.3
